requirements.txt
aiofiles==0.8.0
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
Deprecated==1.2.13
fastapi==0.70.1
h11==0.12.0
httptools==0.3.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pydantic==1.9.0
pyparsing==3.0.6
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
redis==4.1.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==4.0.1
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wrapt==1.13.3
